money_savyy_website/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
yfinance>=0.1.70
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.0.0
scikit-learn>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.10.0
textblob>=0.15.0
gunicorn>=20.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
